Added style applyer and fixed comments
