--- welcome.md
@@ -1,15 +1,0 @@
-# Welcome to OpenCode
-
-## what is open code?
-opencode is an open source text editor made by ivri korem,
-relying on community and the open-source colture. 
-
-## How can i contribute?
-You can create extensions, Editor Themes
-and just improve the source code.
-
-## How can to get started?
-you can go to file menu and create new file,
-and just start coding.
-
-# Happy coding!
